Extracts socket logic to seperate file
